requirements.txt
Flask
requests
beautifulsoup4
pandas
openpyxl
lxml
python-dotenv